apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "momento-api"
version = "0.1.0"
description = "Self-hosted photo management backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.18",
    "pillow>=11.0.0",
    "pillow-heif>=0.21.0",
    "ffmpeg-python>=0.2.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.2",
    "webdavclient3>=3.14.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.14.0",
]

[tool.setuptools]
packages = ["momento_api"]

[tool.black]
line-length = 120
skip-string-normalization = true
skip-magic-trailing-comma = true

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
